[Refactor]#4 remote.service 패키지를 data.service 위치로 이동
