+++ app/src/main/java/com/gallery/kakaogallery/data/repository/ImageRepository.kt
@@ -9,6 +9,8 @@ import com.gallery.kakaogallery.data.entity.remote.request.ImageSearchRequest
 import com.gallery.kakaogallery.data.entity.remote.request.VideoSearchRequest
 import com.gallery.kakaogallery.data.entity.remote.response.ImageSearchModel
 import com.gallery.kakaogallery.data.entity.remote.response.VideoSearchModel
+import com.gallery.kakaogallery.data.service.ImageSearchService
+import com.gallery.kakaogallery.data.service.VideoSearchService
 import com.gallery.kakaogallery.domain.model.ImageModel
 import com.gallery.kakaogallery.domain.model.QuerySearchModel
 import io.reactivex.rxjava3.android.schedulers.AndroidSchedulers
