+++ app/src/main/java/com/gallery/kakaogallery/presentation/viewmodel/SearchImageViewModel.kt
@@ -10,8 +10,8 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import com.gallery.kakaogallery.data.repository.ImageRepository
 import com.gallery.kakaogallery.domain.model.Result
-import com.gallery.kakaogallery.data.remote.service.ImageSearchService
-import com.gallery.kakaogallery.data.remote.service.VideoSearchService
+import com.gallery.kakaogallery.data.service.ImageSearchService
+import com.gallery.kakaogallery.data.service.VideoSearchService
 import com.gallery.kakaogallery.domain.model.ImageModel
 import com.gallery.kakaogallery.data.SaveImageStorage
 import com.gallery.kakaogallery.presentation.application.KakaoGalleryApplication
